+++ bzt/modules/provisioning.py
@@ -27,6 +27,7 @@ class Local(Provisioning):
     """
     Local provisioning means we start all the tools locally
     """
+
     def __init__(self):
         super(Local, self).__init__()
         self.finished_modules = []
@@ -118,3 +119,5 @@ class Local(Provisioning):
             if executor in self.engine.prepared:
                 self.log.debug("Post-process %s", executor)
                 executor.post_process()
+                if executor in self.engine.started and not executor.has_results():
+                    raise RuntimeWarning("Empty results, most likely %s failed" % executor.__class__.__name__)